# specification/swagger.yml
swagger: "2.0"
info:
  description: "Events Planner is a RESTful service designed to create and manage events"
  version: "1.0.0"
  title: "Events Planner"
host: "127.0.0.1:3000"
...
schemes:
- "https"
- "http"
paths:
  /register:
    post:
      tags:
      - "User Actions"
      summary: "Add new user"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User that needs to be added to service"
        required: true
        schema:
          $ref: "#/definitions/UserRegister"
      responses:
        "201":
          description: "Response with newly added user"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Bad Request"
  /login:
    post:
      tags:
      - "User Actions"
      summary: "Login to the platform"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "User that needs to be added to service"
        required: true
        schema:
          $ref: "#/definitions/UserLogin"
      responses:
        "200":
          description: "Response with created user token"
          schema:
            $ref: "#/definitions/UserToken"
        "400":
          description: "Bad Request"
...
definitions:
  UserRegister:
    type: "object"
    properties:
      email:
        type: "string"
      name:
        type: "string"
      surname:
        type: "string"
      password:
        type: "string"
  UserLogin:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
  UserToken:
    type: "object"
    properties:
      token:
        type: "string"
...
  User:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      surname:
        type: "string"
      email:
        type: "string"
...